Moved classes to separate classes module file
